fix: cake product card styling
